+++ app/src/main/java/com/flowfoundation/wallet/page/dialog/processing/send/model/SendProcessingDialogModel.kt
@@ -2,9 +2,10 @@ package com.flowfoundation.wallet.page.dialog.processing.send.model
 
 import com.flowfoundation.wallet.manager.transaction.TransactionState
 import com.flowfoundation.wallet.network.model.UserInfoData
+import java.math.BigDecimal
 
 class SendProcessingDialogModel(
     val userInfo: UserInfoData? = null,
-    val amountConvert: Float? = null,
+    val amountConvert: BigDecimal? = null,
     val stateChange: TransactionState? = null,
 )
