+++ app/src/main/java/com/flowfoundation/wallet/page/send/transaction/subpage/amount/SendAmountViewModel.kt
@@ -13,6 +13,7 @@ import com.flowfoundation.wallet.network.model.AddressBookContact
 import com.flowfoundation.wallet.page.profile.subpage.currency.model.selectedCurrency
 import com.flowfoundation.wallet.page.send.transaction.subpage.amount.model.SendBalanceModel
 import com.flowfoundation.wallet.utils.viewModelIOScope
+import java.math.BigDecimal
 
 class SendAmountViewModel : ViewModel(), OnBalanceUpdate, OnCoinRateUpdate {
     private lateinit var contact: AddressBookContact
@@ -71,7 +72,7 @@ class SendAmountViewModel : ViewModel(), OnBalanceUpdate, OnCoinRateUpdate {
         balanceLiveData.value = data.copy(balance = balance.balance)
     }
 
-    override fun onCoinRateUpdate(coin: FlowCoin, price: Float) {
+    override fun onCoinRateUpdate(coin: FlowCoin, price: BigDecimal) {
         if (coin.contractId() != currentCoin) {
             return
         }
