+++ app/src/main/java/com/flowfoundation/wallet/page/send/transaction/subpage/amount/model/TransactionModel.kt
@@ -5,11 +5,12 @@ import com.google.gson.annotations.SerializedName
 import com.flowfoundation.wallet.manager.coin.FlowCoin
 import com.flowfoundation.wallet.network.model.AddressBookContact
 import kotlinx.parcelize.Parcelize
+import java.math.BigDecimal
 
 @Parcelize
 class TransactionModel(
     @SerializedName("amount")
-    val amount: Float,
+    val amount: BigDecimal,
     @SerializedName("coinId")
     val coinId: String = "",
     @SerializedName("target")
