+++ app/src/main/java/com/flowfoundation/wallet/page/swap/model/SwapModel.kt
@@ -3,6 +3,7 @@ package com.flowfoundation.wallet.page.swap.model
 import com.flowfoundation.wallet.manager.coin.FlowCoin
 import com.flowfoundation.wallet.network.model.SwapEstimateResponse
 import com.google.gson.annotations.SerializedName
+import java.math.BigDecimal
 
 data class SwapModel(
     @SerializedName("fromCoin")
@@ -14,9 +15,9 @@ data class SwapModel(
     @SerializedName("onCoinRateUpdate")
     val onCoinRateUpdate: Boolean? = null,
     @SerializedName("onEstimateFromUpdate")
-    val onEstimateFromUpdate: Float? = null,
+    val onEstimateFromUpdate: BigDecimal? = null,
     @SerializedName("onEstimateToUpdate")
-    val onEstimateToUpdate: Float? = null,
+    val onEstimateToUpdate: BigDecimal? = null,
     @SerializedName("onEstimateLoading")
     val onEstimateLoading: Boolean? = null,
     @SerializedName("estimateData")
